Increase robustness and allow multiple grammars.
The base app takes a sub-url which become the name of the grammar, so .../demophin/erg/ is for the "erg" grammar. These grammar names are defined in the config file, with paths to the grammar images, descriptions, etc. This allows one Demophin instance to handle multiple grammars.

For convenience, the .../parse URL can also be used with GET, like this:
  ...demophin/erg/parse?sentence=Dogs bark

This commit also adds a bit more error handling.